# openapi.yml
...
openapi: 3.0.0
paths:
...
  /settings/team:
    get:
      operationId: SettingsController_getTeam
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - Settings
    post:
      operationId: SettingsController_inviteTeamMember
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTeamMemberDto"
      responses:
        "201":
          description: ""
      tags:
        - Settings
  /settings/team/{id}/role:
    put:
      operationId: SettingsController_changeTeamMemberRole
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Settings
  /settings/team/create-user:
    post:
      operationId: SettingsController_createTeamUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamUserDto"
      responses:
        "201":
          description: ""
      tags:
        - Settings
  /settings/team/{id}:
    delete:
      operationId: SettingsController_deleteTeamMember
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Settings
  /settings/shortlink:
    get:
      operationId: SettingsController_getShortlinkPreference
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - Settings
    post:
      operationId: SettingsController_updateShortlinkPreference
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShortlinkPreferenceDto"
      responses:
        "201":
          description: ""
      tags:
        - Settings
...
  /admin/ai-settings/rag/backfill:
    post:
      operationId: AiSettingsController_triggerRagBackfill
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerRagBackfillDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
      tags:
        - AI Settings
...
  /rag/items:
    get:
      operationId: RagController_listItems
      parameters:
        - name: offset
          required: false
          in: query
          schema:
            type: string
        - name: limit
          required: false
          in: query
          schema:
            type: string
        - name: sourceType
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - RAG
  /rag/items/{sourceType}/{sourceId}:
    delete:
      operationId: RagController_deleteItem
      parameters:
        - name: sourceType
          required: true
          in: path
          schema:
            type: string
        - name: sourceId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - RAG
  /rag/search:
    post:
      operationId: RagController_search
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RagSearchDto"
      responses:
        "201":
          description: ""
      tags:
        - RAG
...
info:
  title: ValidPost API
  description: OpenAPI description of the ValidPost backend (NestJS). Generated from the controllers — do not hand-edit; run `pnpm run openapi:generate`.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    api-key:
      type: apiKey
      in: header
      name: Authorization
    bearer:
      scheme: bearer
      bearerFormat: Token
      type: http
  schemas:
...
    AddTeamMemberDto:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          deprecated: true
          enum:
            - USER
            - ADMIN
        roleId:
          type: string
        sendEmail:
          type: boolean
      required:
        - email
        - sendEmail
    CreateTeamUserDto:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          deprecated: true
          default: USER
          enum:
            - USER
            - ADMIN
        roleId:
          type: string
      required:
        - email
        - password
        - role
    ShortlinkPreferenceDto:
      type: object
      properties:
        shortlink:
          type: string
          enum:
            - ASK
            - YES
            - NO
      required:
        - shortlink
...
    TriggerRagBackfillDto:
      type: object
      properties: {}
...
    RagSearchDto:
      type: object
      properties:
        query:
          type: string
          minLength: 1
        limit:
          type: number
          minimum: 1
      required:
        - query
...